Extract footer text to function
