+++ john-macgregor/src/components/app/appFunctions.js
@@ -1,0 +1,8 @@
+const functions = {
+    footerText: () => {
+        const year = new Date().getFullYear();
+        return `${year} Johnathan MacGregor`;
+    }
+}
+
+export default functions;
